Separate thread and call modules; additional refactoring
