+++ src/prefect/_internal/concurrency/calls.py
@@ -1,18 +1,15 @@
 import abc
 import asyncio
-import atexit
 import concurrent.futures
 import contextlib
 import contextvars
 import dataclasses
 import inspect
-import threading
 from typing import Any, Callable, Dict, Generic, Optional, Tuple, TypeVar
 
 from typing_extensions import ParamSpec
 
 from prefect._internal.concurrency.event_loop import get_running_loop
-from prefect._internal.concurrency.primitives import Event
 from prefect._internal.concurrency.timeouts import (
     cancel_async_at,
     cancel_sync_at,
@@ -24,7 +21,7 @@ T = TypeVar("T")
 P = ParamSpec("P")
 
 
-logger = get_logger("prefect._internal.concurrency.portals")
+logger = get_logger("prefect._internal.concurrency.calls")
 
 
 # Tracks the current call being executed
@@ -57,7 +54,7 @@ class Call(Generic[T]):
     context: contextvars.Context
     deadline: Optional[float] = None
     portal: Optional["Portal"] = None
-    callback_handler: Optional[Callable[["Call"], None]] = None
+    callback_portal: Optional["Portal"] = None
 
     @classmethod
     def new(cls, __fn: Callable[P, T], *args: P.args, **kwargs: P.kwargs) -> "Call[T]":
@@ -89,23 +86,23 @@ class Call(Generic[T]):
 
         self.portal = portal
 
-    def set_callback_handler(self, handler: Callable[["Call"], None]) -> None:
+    def set_callback_portal(self, portal: "Portal") -> None:
         """
-        Set the callback handler for this call.
+        Set a portal to handle callbacks for this call.
         """
-        if self.callback_handler is not None:
-            raise RuntimeError("A callback handler has already been set for this call.")
+        if self.callback_portal is not None:
+            raise RuntimeError("A callback portal has already been set for this call.")
 
-        self.callback_handler = handler
+        self.callback_portal = portal
 
     def add_callback(self, call: "Call") -> None:
         """
         Send a callback to the callback handler.
         """
-        if self.callback_handler is None:
+        if self.callback_portal is None:
             raise RuntimeError("No callback handler has been configured.")
 
-        self.callback_handler(call)
+        self.callback_portal.submit(call)
 
     def run(self) -> None:
         """
@@ -213,140 +210,11 @@ class Portal(abc.ABC):
     """
 
     @abc.abstractmethod
-    def start(self):
-        """
-        Start the portal.
-        """
-
-    @abc.abstractmethod
     def submit(self, call: "Call") -> "Call":
         """
         Submit a call to execute elsewhere.
 
         The call's result can be retrieved with `call.result()`.
 
-        Returns the call.
+        Returns the call for convenience.
         """
-
-    @abc.abstractmethod
-    def shutdown(self) -> None:
-        """
-        Shutdown the portal.
-        """
-
-    @abc.abstractproperty
-    def name(self) -> str:
-        """
-        Get the name of the portal.
-        """
-
-    def __enter__(self):
-        self.start()
-        return self
-
-    def __exit__(self, *_):
-        self.shutdown()
-
-
-class WorkerThreadPortal(Portal):
-    """
-    A portal to a worker running on a thread with an event loop.
-    """
-
-    def __init__(
-        self, name: str = "WorkerThread", daemon: bool = False, run_once: bool = False
-    ):
-        self.thread = threading.Thread(
-            name=name, daemon=daemon, target=self._entrypoint
-        )
-        self._ready_future = concurrent.futures.Future()
-        self._loop: Optional[asyncio.AbstractEventLoop] = None
-        self._shutdown_event: Event = Event()
-        self._run_once: bool = run_once
-        self._submitted_count: int = 0
-
-        if not daemon:
-            atexit.register(self.shutdown)
-
-    def start(self):
-        """
-        Start the worker thread; raises any exceptions encountered during startup.
-        """
-        self.thread.start()
-        # Wait for the worker to be ready
-        self._ready_future.result()
-
-    def submit(self, call: Call) -> Call:
-        if self._submitted_count > 0 and self._run_once:
-            raise RuntimeError(
-                "Worker configured to only run once. A call has already been submitted."
-            )
-
-        if self._loop is None:
-            self.start()
-
-        if self._shutdown_event.is_set():
-            raise RuntimeError("Worker is shutdown.")
-
-        # Track the portal running the call
-        call.set_portal(self)
-
-        self._loop.call_soon_threadsafe(call.run)
-        self._submitted_count += 1
-        if self._run_once:
-            call.future.add_done_callback(lambda _: self.shutdown())
-
-        return call
-
-    def shutdown(self) -> None:
-        """
-        Shutdown the worker thread. Does not wait for the thread to stop.
-        """
-        if not self._shutdown_event:
-            return
-
-        self._shutdown_event.set()
-        # TODO: Consider blocking on `thread.join` here?
-
-    @property
-    def name(self) -> str:
-        return self.thread.name
-
-    def _entrypoint(self):
-        """
-        Entrypoint for the thread.
-
-        Immediately create a new event loop and pass control to `run_until_shutdown`.
-        """
-        try:
-            asyncio.run(self._run_until_shutdown())
-        except BaseException:
-            # Log exceptions that crash the thread
-            logger.exception("%s encountered exception", self.name)
-            raise
-
-    async def _run_until_shutdown(self):
-        try:
-            self._loop = asyncio.get_running_loop()
-            self._ready_future.set_result(True)
-        except Exception as exc:
-            self._ready_future.set_exception(exc)
-            return
-
-        await self._shutdown_event.wait()
-
-
-GLOBAL_THREAD_PORTAL: Optional[WorkerThreadPortal] = None
-
-
-def get_global_thread_portal() -> WorkerThreadPortal:
-    global GLOBAL_THREAD_PORTAL
-
-    # Create a new worker on first call or if the existing worker is dead
-    if GLOBAL_THREAD_PORTAL is None or not GLOBAL_THREAD_PORTAL.thread.is_alive():
-        GLOBAL_THREAD_PORTAL = WorkerThreadPortal(
-            daemon=True, name="GlobalWorkerThread"
-        )
-        GLOBAL_THREAD_PORTAL.start()
-
-    return GLOBAL_THREAD_PORTAL
